+++ src/main/java/com/cryptowallet/controller/UserController.java
@@ -3,6 +3,7 @@ package com.cryptowallet.controller;
 import com.cryptowallet.entities.User;
 import com.cryptowallet.services.RoleService;
 import com.cryptowallet.services.UserService;
+import com.cryptowallet.services.facades.UserServiceFacade;
 import com.cryptowallet.utils.PasswordGenerator;
 import com.cryptowallet.utils.UsersRoles;
 import com.cryptowallet.utils.ValidateInputData;
@@ -12,9 +13,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -23,18 +22,20 @@ import java.util.UUID;
 
 @Controller
 public class UserController {
-    public static final int TOKEN_LENGTH = 50;
+    private final String VERIFIED = "Your account is verified!";
+    private final String NOT_VERIFIED = "Your account is not verified!";
+    private final String ACTIVATION_MESSAGE = "A letter was sent to your mail. " +
+            "Please follow the link in the letter to create a new password";
+    private final String PASSWORD_SAVED ="Your password has been successfully saved!";
 
+    private UserServiceFacade userServiceFacade;
     private UserService userService;
-    private RoleService roleService;
-    private PasswordEncoder passwordEncoder;
     private ValidateInputData validError;
 
     @Autowired
-    public UserController(UserService userService, RoleService roleService, PasswordEncoder passwordEncoder) {
+    public UserController(UserService userService, UserServiceFacade userServiceFacade) {
         this.userService = userService;
-        this.roleService = roleService;
-        this.passwordEncoder = passwordEncoder;
+        this.userServiceFacade = userServiceFacade;
         this.validError = new ValidateInputData();
     }
 
@@ -54,11 +55,10 @@ public class UserController {
                 model.addAttribute("not_valid", validError.getValidationErrors());
                 return "registration";
             }
-            user.setPassword(passwordEncoder.encode(password));
-            user.setRole(roleService.getRoleById(UsersRoles.ROLE_USER.getRole()).get());
-            user.setToken(PasswordGenerator.generate(TOKEN_LENGTH));
+            user.setPassword(userServiceFacade.passwordEncode(password));
+            user.setRole(userServiceFacade.getUserRole());
+            user.setToken(userServiceFacade.generateToken());
             userService.saveUser(user);
-
         } else {
             validError.putValidationErrors("User already exists", "User already exists");
             user.setEmail("");
@@ -79,6 +79,80 @@ public class UserController {
     @GetMapping("/login")
     public String getLoginPage() {
         return "login";
+    }
+
+    @GetMapping("/activate/{code}")
+    public String activateUser (Model model, @PathVariable String code)  {
+        User user = userService.findByToken(code);
+        if(user!=null) {
+            user.setApproved(true);
+            userService.saveUser(user);
+            model.addAttribute("activeMessage", VERIFIED);
+            return "index";
+        }else {
+            model.addAttribute("activeMessage", NOT_VERIFIED);
+        return "index";
+        }
+    }
+
+    @GetMapping("/restorePassword")
+    public String getRestorePassword() {
+        return "restorePassword";
+    }
+
+    @PostMapping("/restorePassword")
+    public String sendMailRestorePassword (Model model, @RequestParam(name = "email") String email) {
+        validError.clearValidate();
+        if(validError.isEmailValid(email)) {
+            if (userService.isUserExist(email)){
+                userServiceFacade.restorePassword(email);
+                model.addAttribute("activeMessage", ACTIVATION_MESSAGE);
+                return "index";
+            }else {
+                validError.putValidationErrors("User dont exist", "User dont exist");
+                model.addAttribute("not_valid", validError.getValidationErrors());
+                return "restorePassword";
+            }
+        }else {
+            model.addAttribute("not_valid", validError.getValidationErrors());
+            return "restorePassword";
+        }
+    }
+
+    @GetMapping("/restore/{code}")
+    public String restoreUser (Model model, @PathVariable String code) {
+        User user = userService.findByToken(code);
+        if(user!=null) {
+            model.addAttribute("email", user.getEmail());
+            return "newPassword";
+        }else {
+            model.addAttribute("activeMessage", NOT_VERIFIED);
+        }
+        return "index";
+    }
+
+    @PostMapping("/newPassword")
+    public String saveNewPassword (Model model, @RequestParam(name = "password") String password,
+                                   @RequestParam(name = "login") String login) {
+        User user = userService.findByLogin(login);
+        if (user!=null){
+            validError.clearValidate();
+            if (!validError.isPasswordValid(password)) {
+                model.addAttribute("login", login);
+                model.addAttribute("not_valid", validError.getValidationErrors().values());
+                return "newPassword";
+            }else {
+                user.setPassword(userServiceFacade.passwordEncode(password));
+                userService.saveUser(user);
+                model.addAttribute("activeMessage", PASSWORD_SAVED);
+                return "index";
+            }
+        }else {
+            model.addAttribute("login", login);
+            validError.putValidationErrors("User dont exist", "User dont exist");
+            model.addAttribute("not_valid", validError.getValidationErrors());
+            return "newPassword";
+        }
     }
 
 }
